+++ Changelog.md
@@ -1,6 +1,12 @@
 ### Development
 [Full Changelog](http://github.com/rspec/rspec-support/compare/v3.5.0.beta1...master)
 
+Bug Fixes:
+
+* Fix `ObjectFormatter` so that formatting objects that don't respond to
+  `#inspect` (such as `BasicObject`) does not cause `NoMethodError`.
+  (Yuji Nakayama, #269)
+
 ### 3.5.0.beta1 / 2016-02-06
 [Full Changelog](http://github.com/rspec/rspec-support/compare/v3.4.1...v3.5.0.beta1)